few more cleanups
