+++ go/common/backup/pgbackrest_repo_test.go
@@ -18,9 +18,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 )
 
 func TestInitialPgBackRestRepo(t *testing.T) {
@@ -58,55 +55,4 @@ func TestInitialPgBackRestRepo(t *testing.T) {
 			assert.Equal(t, tt.want, InitialPgBackRestRepo(tt.keys))
 		})
 	}
-}
-
-func TestRepoStorageConfigGenerationPaths(t *testing.T) {
-	fsCfg, err := NewConfig(&clustermetadatapb.BackupLocation{
-		Location: &clustermetadatapb.BackupLocation_Filesystem{
-			Filesystem: &clustermetadatapb.FilesystemBackup{Path: "/backups"},
-		},
-	})
-	require.NoError(t, err)
-	s3Cfg, err := NewConfig(&clustermetadatapb.BackupLocation{
-		Location: &clustermetadatapb.BackupLocation_S3{
-			S3: &clustermetadatapb.S3Backup{Bucket: "bucket", Region: "us-east-1", KeyPrefix: "prod/"},
-		},
-	})
-	require.NoError(t, err)
-
-	// Generation 1 is the base path (byte-identical to the pre-generation
-	// layout); later generations get a distinct gen-<N> component.
-	cfg, err := fsCfg.PgBackRestConfig(1, 1, "multigres")
-	require.NoError(t, err)
-	assert.Equal(t, "/backups", cfg["repo1-path"])
-
-	cfg, err = fsCfg.PgBackRestConfig(2, 3, "multigres")
-	require.NoError(t, err)
-	assert.Equal(t, "/backups/gen-3", cfg["repo2-path"])
-	assert.Equal(t, "posix", cfg["repo2-type"])
-
-	cfg, err = s3Cfg.PgBackRestConfig(1, 1, "multigres")
-	require.NoError(t, err)
-	assert.Equal(t, "/prod/multigres", cfg["repo1-path"])
-
-	cfg, err = s3Cfg.PgBackRestConfig(2, 2, "multigres")
-	require.NoError(t, err)
-	assert.Equal(t, "/prod/multigres/gen-2", cfg["repo2-path"])
-	assert.Equal(t, "bucket", cfg["repo2-s3-bucket"])
-}
-
-func TestRepoRetentionConfig(t *testing.T) {
-	// repo1 carries the default retention values.
-	cfg := repoRetentionConfig(1)
-	assert.Equal(t, RetentionDifferential, cfg["repo1-retention-diff"])
-	assert.Equal(t, RetentionFull, cfg["repo1-retention-full"])
-	assert.Equal(t, RetentionFullType, cfg["repo1-retention-full-type"])
-	assert.Equal(t, RetentionHistory, cfg["repo1-retention-history"])
-
-	// A later repo index prefixes every key with its own repoN-.
-	cfg = repoRetentionConfig(2)
-	assert.Equal(t, RetentionDifferential, cfg["repo2-retention-diff"])
-	assert.Equal(t, RetentionFull, cfg["repo2-retention-full"])
-	assert.Equal(t, RetentionFullType, cfg["repo2-retention-full-type"])
-	assert.Equal(t, RetentionHistory, cfg["repo2-retention-history"])
 }
